Added permutations and automatic data creation.
Created the permutations module written in C. This is located in the src
subdirectory of the main branch. Other additional modules to be written
in C should be located there, for instance rational numbers and
matrices. I have not yet written an API for the CPython module.

Other major contribution is the automatic calculation of data for the
exceptional type Weyl groups. This relies on some computations performed
with GAP4, specifically for the calculation of the ordinary character
tables. A script is provided to automate the calculation of the raw
data, which is then put into the file data/rawdata.py.