Generalised CoxeterCoset to general automorphisms
The previous version of CoxeterCoset would only accept automorphisms
which came from a permutation of the roots. It now accepts arbitrary
Coxeter automorphisms, i.e., those permuting the roots up to scalar
multiples. This allows one to deal with cases such as 2B2, 2G2, 2F4. The
resulting group is represented as a permutation on the 1-dimensional
lines spanned by the roots.